% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy_procs_Oct26.R
\name{ar_sim}
\alias{ar_sim}
\title{Simulate many AR(1) processes with zero mean}
\usage{
ar_sim(a = 0.5, n_sample = 500, n_mc = 1000)
}
\arguments{
\item{a}{AR parameter}

\item{n_sample}{Sample size}

\item{n_mc}{Nr of MC iterations}
}
\value{
matrix, each col is a simulated AR process
}
\description{
Simulate many AR(1) processes with zero mean
}
