% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy_procs_Jan22.R
\name{ptest_VaRES}
\alias{ptest_VaRES}
\alias{ptest_VaRES_full_2_boot}
\title{Westfall-young procedure for VaR-ES forecasts}
\usage{
ptest_VaRES(TSfc, alpha = 0.025, np = 500, grid_thin = 1,
  block_length = 0.7343665 * (nrow(TSfc)^(1/3)), HAC_lags = 1,
  score_type = "2", grid_bound = NA, sampling_method = "bootstrap",
  small_output = FALSE)

ptest_VaRES_full_2_boot(TSfc, alpha = 0.025, np = 500, block_length = 10 *
  (nrow(TSfc)/2525)^(1/3))
}
\arguments{
\item{TSfc}{Matrix containing forecasts and realizations. First two columns contain VaR/ES forecasts from method A; cols 3-4 contain VaR/ES forecasts from method B; col 5 contains realizations.}

\item{alpha}{Scalar, level of VaR/ES to be predicted}

\item{np}{Nr of iterations (for bootstrap or permutation)}

\item{grid_thin}{Thinning of grid (1 means no thinning, i.e. all forecasts in sample used as grid points)}

\item{block_length}{Mean block length for bootstrap or permutation}

\item{HAC_lags}{Nr of lags for HAC estimator}

\item{score_type}{Either "12" (both scores) or "2" (second score only)}

\item{grid_bound}{Upper bound for jumps (only needed if grid_thin > 1)}

\item{sampling_method}{Either "bootstrap" or "permutation"}

\item{small_output}{Dummy, set to TRUE to reduce list of outputs (e.g. for use in simulations)}
}
\value{
\code{pval_WY} is the p-value of the full Westfall-Young procedure; \code{pval_OS} is the p-value of the simplified one-step procedure (which is equivalent to Hansen's test)
}
\description{
Westfall-young procedure for VaR-ES forecasts
}
\details{
Tests the hypothesis that the first method weakly dominates the second method. \code{ptest_VaRES} employs a finite grid approximation, whereas \code{ptest_VaRES_full_2_boot} uses analytical calculations. See Section 3.2 in the paper for details.
}
\author{
Alexander Jordan, Fabian Krueger, Johanna Ziegel
}
