% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy_procs_Jan22.R
\name{sim_VaRES}
\alias{sim_VaRES}
\title{Simulate data from the design of Section 4 in the paper.}
\usage{
sim_VaRES(T, s2_0 = 0.35, a = 0.5, b = 0.7, zeta1 = 1, zeta2 = 1,
  t_df = 6, alpha = 0.025)
}
\arguments{
\item{T}{Length of series to be simulated.}

\item{s2_0, a, b, t_df}{Parameters of time series model which generates 'true' forecasts}

\item{zeta1, zeta2}{Variances of disturbances}

\item{alpha}{Level of VaR/ES forecasts}
}
\description{
Simulate data from the design of Section 4 in the paper.
}
\details{
Simulate time series of VaR/ES forecasts and corresponding realizations, as described in the paper by Ziegel et al (2018).
}
