% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murphy_procs_Oct26.R
\name{stationary_boot}
\alias{stationary_boot}
\title{T-tests using stationary bootstrap}
\usage{
stationary_boot(dat, n_boot, block_length, eps = 1e-06)
}
\arguments{
\item{dat}{Matrix, each column is a time series}

\item{n_boot}{Nr of bootstrap iterations}

\item{block_length}{Expected block length of stationary bootstrap}

\item{eps}{Threshold for setting t-stat to NA}
}
\value{
t-stats
}
\description{
T-tests using stationary bootstrap
}
